v01 Warmstate function updates (#142)
* remove extraneous function inputs

* Add `=None` default value to optional filter args

* Clarify docstrings for warmstates functions

* add function for csv-based qlat input

Also improve docstring

* move io for qlat, channel, waterbody to nhd_io

* move waterbody read functions to v02 nhd_io

* remove extra read_level_pool_waterbody_df_v02 function

was named read_waterbody_df in the nhd_io.py file prior to this PR.

* remove extra sort

* convert read_dataset to `with` style

* convert all read_dataset calls to `with` style

* move to a dataset filter in read_waterbody_df
