+++ src/python_framework_v02/nhd_io.py
@@ -6,7 +6,8 @@ import geopandas as gpd
 
 
 def read_netcdf(geo_file_path):
-    return xr.open_dataset(geo_file_path).to_dataframe()
+    with xr.open_dataset(geo_file_path) as ds:
+        return ds.to_dataframe()
 
 
 def read_csv(geo_file_path, header="infer", layer_string=None):
@@ -37,14 +38,6 @@ def read_mask(path, layer_string=None):
     return read_csv(path, header=None, layer_string=layer_string)
 
 
-def read_qlat(path):
-    ql = pd.read_csv(path, index_col=0)
-    ql.index = ql.index.astype(int)
-    ql.columns = ql.columns.astype(int)
-    ql = ql.sort_index(axis="index")
-    return ql.astype("float32")
-
-
 def replace_downstreams(data, downstream_col, terminal_code):
     ds0_mask = data[downstream_col] == terminal_code
     new_data = data.copy()
@@ -55,9 +48,211 @@ def replace_downstreams(data, downstream_col, terminal_code):
     return new_data
 
 
-def read_waterbody_df(parm_file, lake_id_mask=None):
-    df1 = xr.open_dataset(parm_file).to_dataframe()
-    df1 = df1.set_index("lake_id").sort_index()
+def read_waterbody_df(waterbody_parameters, waterbodies_values, wbtype="level_pool"):
+    """
+    General waterbody dataframe reader. At present, only level-pool
+    capability exists. 
+    """
+    if wbtype == "level_pool":
+        wb_params = waterbody_parameters[wbtype]
+        return read_level_pool_waterbody_df(
+            wb_params["level_pool_waterbody_parameter_file_path"],
+            wb_params["level_pool_waterbody_id"],
+            waterbodies_values[wbtype],
+        )
+
+
+def read_level_pool_waterbody_df(
+    parm_file, lake_index_field="lake_id", lake_id_mask=None
+):
+    """
+    Reads LAKEPARM file and prepares a dataframe, filtered 
+    to the relevant reservoirs, to provide the parameters
+    for level-pool reservoir computation.
+
+    Completely replaces the read_waterbody_df function from prior versions
+    of the v02 routing code.
+
+    Prior version filtered the dataframe as opposed to the dataset as in this version. 
+    with xr.open_dataset(parm_file) as ds:
+        df1 = ds.to_dataframe()
+    df1 = df1.set_index(lake_index_field).sort_index(axis="index")
     if lake_id_mask is None:
         return df1
-    return df1.loc[lake_id_mask]
+    else:
+        return df1.loc[lake_id_mask]
+    """
+
+    #TODO: avoid or parameterize "feature_id" or ... return to name-blind dataframe version
+    with xr.open_dataset(parm_file) as ds:
+        ds = ds.swap_dims({"feature_id": lake_index_field})
+        df1 = ds.sel({lake_index_field: list(lake_id_mask)}).to_dataframe()
+    df1 = df1.sort_index(axis="index")
+    return df1
+
+
+def get_ql_from_csv(qlat_input_file, index_col=0):
+    """
+    qlat_input_file: comma delimted file with header giving timesteps, rows for each segment
+    index_col = 0: column/field in the input file with the segment/link id
+    """
+    ql = pd.read_csv(qlat_input_file, index_col=index_col)
+    ql.index = ql.index.astype(int)
+    ql.columns = ql.columns.astype(int)
+    ql = ql.sort_index(axis="index")
+    return ql.astype("float32")
+
+
+def read_qlat(path):
+    """
+    retained for backwards compatibility with early v02 files
+    """
+    return get_ql_from_csv(path)
+
+
+def get_ql_from_wrf_hydro(qlat_files, index_col="station_id", value_col="q_lateral"):
+    """
+    qlat_files: globbed list of CHRTOUT files containing desired lateral inflows
+    index_col: column/field in the CHRTOUT files with the segment/link id
+    value_col: column/field in the CHRTOUT files with the lateral inflow value
+
+    In general the CHRTOUT files contain one value per time step. At present, there is
+    no capability for handling non-uniform timesteps in the qlaterals.
+
+    The qlateral may also be input using comma delimited file -- see
+    `get_ql_from_csv`
+    """
+
+    li = []
+
+    for filename in qlat_files:
+        with xr.open_dataset(filename) as ds:
+            df1 = ds.to_dataframe()
+
+        li.append(df1)
+
+    frame = pd.concat(li, axis=0, ignore_index=False)
+    mod = frame.reset_index()
+    ql = mod.pivot(index=index_col, columns="time", values=value_col)
+
+    return ql
+
+
+def get_stream_restart_from_wrf_hydro(
+    channel_initial_states_file,
+    crosswalk_file,
+    channel_ID_column,
+    us_flow_column="qlink1",
+    ds_flow_column="qlink2",
+    depth_column="hlink",
+    default_us_flow_column="qu0",
+    default_ds_flow_column="qd0",
+    default_depth_column="h0",
+):
+    """
+    channel_initial_states_file: WRF-HYDRO standard restart file
+    crosswalk_file: File containing channel IDs IN THE ORDER of the Restart File
+    channel_ID_column: field in the crosswalk file to assign as the index of the restart values
+    us_flow_column: column in the restart file to use for upstream flow initial state
+    ds_flow_column: column in the restart file to use for downstream flow initial state
+    depth_column: column in the restart file to use for depth initial state
+    default_us_flow_column: name used in remainder of program to refer to this column of the dataset
+    default_ds_flow_column: name used in remainder of program to refer to this column of the dataset
+    default_depth_column: name used in remainder of program to refer to this column of the dataset
+
+    The Restart file gives hlink, qlink1, and qlink2 values for channels --
+    the order is simply the same as that found in the Route-Link files.
+    *Subnote 1*: The order of these values is NOT the order found in the CHRTOUT files,
+    though the number of outputs is the same as in those files.
+    *Subnote 2*: If there is no mask applied, then providing the path to the RouteLink file should
+    be sufficient for the crosswalk file. If there is a mask applied (so that
+    not all segments in the RouteLink file are used in the routing computations) then
+    a pre-processing step will be need to provide only the relevant segments in the
+    crosswalk file.
+    """
+
+    with xr.open_dataset(crosswalk_file) as xds:
+        xdf = xds[channel_ID_column].to_dataframe()
+    xdf = xdf.reset_index()
+    xdf = xdf[[channel_ID_column]]
+    with xr.open_dataset(channel_initial_states_file) as qds:
+        if depth_column in qds:
+            qdf2 = qds[[us_flow_column, ds_flow_column, depth_column]].to_dataframe()
+        else:
+            qdf2 = qds[[us_flow_column, ds_flow_column]].to_dataframe()
+            qdf2[depth_column] = 0
+    qdf2 = qdf2.reset_index()
+    qdf2 = qdf2[[us_flow_column, ds_flow_column, depth_column]]
+    qdf2.rename(
+        columns={
+            us_flow_column: default_us_flow_column,
+            ds_flow_column: default_ds_flow_column,
+            depth_column: default_depth_column,
+        },
+        inplace=True,
+    )
+    qdf2[channel_ID_column] = xdf
+    qdf2 = qdf2.reset_index().set_index([channel_ID_column])
+
+    q_initial_states = qdf2
+
+    return q_initial_states
+
+
+def get_reservoir_restart_from_wrf_hydro(
+    waterbody_intial_states_file,
+    crosswalk_file,
+    waterbody_ID_field,
+    crosswalk_filter_file=None,
+    crosswalk_filter_file_field=None,
+    waterbody_flow_column="qlakeo",
+    waterbody_depth_column="resht",
+    default_waterbody_flow_column="qd0",
+    default_waterbody_depth_column="h0",
+):
+    """
+    waterbody_intial_states_file: WRF-HYDRO standard restart file
+    crosswalk_file: File containing reservoir IDs IN THE ORDER of the Restart File
+    waterbody_ID_field: field in the crosswalk file to assign as the index of the restart values
+    crosswalk_filter_file: file containing only reservoir ids needed in the crosswalk file (see notes below)
+    crosswalk_filter_file_field: waterbody id field in crosswalk filter file
+    waterbody_flow_column: column in the restart file to use for upstream flow initial state
+    waterbody_depth_column: column in the restart file to use for downstream flow initial state
+    default_waterbody_flow_column: name used in remainder of program to refer to this column of the dataset
+    default_waterbody_depth_column: name used in remainder of program to refer to this column of the dataset
+
+    The Restart file gives qlakeo and resht values for waterbodies.
+    The order of values in the file is the same as the order in the LAKEPARM file from WRF-Hydro.
+    However, there are many instances where only a subset of waterbodies described in the lakeparm
+    file are used. In these cases, a filter file must be provided which specifies
+    which of the reservoirs in the crosswalk file are to be used.
+    """
+
+    with xr.open_dataset(crosswalk_file) as xds:
+        X = xds[waterbody_ID_field]
+
+        if crosswalk_filter_file:
+            with xr.open_dataset(crosswalk_filter_file) as fds:
+                xdf = X.loc[X.isin(fds[crosswalk_filter_file_field])].to_dataframe()
+        else:
+            xdf = X.to_dataframe()
+
+    xdf = xdf.reset_index()[waterbody_ID_field]
+
+    # read initial states from r&r output
+    with xr.open_dataset(waterbody_intial_states_file) as resds:
+        resdf = resds[[waterbody_flow_column, waterbody_depth_column]].to_dataframe()
+    resdf = resdf.reset_index()
+    resdf = resdf[[waterbody_flow_column, waterbody_depth_column]]
+    resdf.rename(
+        columns={
+            waterbody_flow_column: default_waterbody_flow_column,
+            waterbody_depth_column: default_waterbody_depth_column,
+        },
+        inplace=True,
+    )
+
+    mod = resdf.join(xdf).reset_index().set_index(waterbody_ID_field)
+    init_waterbody_states = mod
+
+    return init_waterbody_states
